POType -> POTextType
