+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/style/text/POTextStyle.kt
@@ -8,5 +8,5 @@ import kotlinx.parcelize.Parcelize
 data class POTextStyle(
     @ColorRes
     val colorResId: Int,
-    val type: POType
+    val type: POTextType
 ) : Parcelable
